Move inline/* to *-inl.h
