+++ include/simdjson/dom/document_stream-inl.h
